update model tests
