+++ tests/test_cmip6_dataset_model.py
@@ -5,21 +5,12 @@ import pytest
 from pydantic import ValidationError
 
 from piddiplatsch.config import config
-from piddiplatsch.models import CMIP6DatasetModel, CMIP6FileModel, HostingNode
+from piddiplatsch.models import CMIP6DatasetModel, HostingNode
 
 # --- Helper constants ---
 HOST_NODE = HostingNode(
     host="esgf-node", published_on=datetime.datetime.now(datetime.UTC)
 )
-VALID_FILE_DATA = {
-    "URL": "https://example.com/data.nc",
-    "FILE_NAME": "data.nc",
-    "IS_PART_OF": "dataset-001",
-    "CHECKSUM": "12205994471abb01112afcc18159f6cc74b4f511b99806da59b3caf5a9c173cacfc5",
-    "CHECKSUM_METHOD": "sha2-256",
-    "FILE_SIZE": 123,
-    "DOWNLOAD_URL": "https://example.com/data.nc",
-}
 
 
 # --- CMIP6DatasetModel tests ---
@@ -141,23 +132,3 @@ def test_optional_fields_defaults():
     assert dataset._RETRACTED is False
 
 
-# --- CMIP6FileModel tests ---
-def test_file_model_checksum_validation():
-    config._set("schema", "strict_mode", True)
-
-    file_model = CMIP6FileModel(**VALID_FILE_DATA)
-    assert file_model.CHECKSUM_METHOD == "sha2-256"
-
-    data = VALID_FILE_DATA.copy()
-    data["CHECKSUM"] = "invalidchecksum"
-    data["CHECKSUM_METHOD"] = "unknown"
-    with pytest.raises(ValueError):
-        CMIP6FileModel(**data)
-
-
-def test_file_model_url_serialization():
-    file_model = CMIP6FileModel(**VALID_FILE_DATA)
-    data = file_model.model_dump()  # triggers field_serializers
-    assert isinstance(data["DOWNLOAD_URL"], str)
-    for url in data["REPLICA_DOWNLOAD_URLS"]:
-        assert isinstance(url, str)
